Moving table validation to utilities
