+++ urbansim_templates/data/load_table.py
@@ -221,78 +221,3 @@ class LoadTable():
                 return df
             
         
-    def validate(self):
-        """
-        Check some basic expectations about the table generated by the step:
-        
-        - Confirm that the table includes a unique, named index column (primary key) or 
-          set of columns (composite key). If not, raise a ValueError.
-        
-        - If the table contains columns whose names match the index columns of tables
-          previously registered with Orca, check whether they make sense as join keys.
-          Print a status message with the number of presumptive foreign-key values that 
-          are found in the primary key column. 
-        
-        - Perform the same check for columns in previously registered tables whose names
-          match the index of the table generated by this step.
-          
-        - It doesn't currently compare indexes to indexes. (Maybe it should?)
-          
-        Running this will trigger loading all registered Orca tables into memory, which 
-        may take a while if they have not yet been loaded. Stand-alone columns will not 
-        be loaded unless their names match an index column. 
-        
-        Returns
-        -------
-        bool
-        
-        """
-        # There are a couple of reasons we're not using the orca_test library here:
-        # (a) orca_test doesn't currently support MultiIndexes, and (b) the primary-key/
-        # foreign-key comparisons aren't asserting anything, just printing status 
-        # messages. We should update orca_test to support both, probably.
-        
-        # Register table if needed
-        if not orca.is_table(self.table):
-            self.run()
-        
-        idx = orca.get_table(self.table).index
-        
-        # Check index has a name
-        if list(idx.names) == [None]:
-            raise ValueError("Index column has no name")
-        
-        # Check index is unique
-        if len(idx.unique()) < len(idx):    
-            raise ValueError("Index not unique")
-        
-        # Compare columns to indexes of other tables, and vice versa
-        combinations = [(self.table, t) for t in orca.list_tables() if self.table != t] \
-                + [(t, self.table) for t in orca.list_tables() if self.table != t]
-        
-        for t1, t2 in combinations:
-            col_names = orca.get_table(t1).columns
-            idx = orca.get_table(t2).index
-            
-            if set(idx.names).issubset(col_names):
-                vals = orca.get_table(t1).to_frame(idx.names).drop_duplicates()
-                
-                # Easier to compare multi-column values to multi-column index if we 
-                # turn the values into an index as well
-                vals = vals.reset_index().set_index(idx.names).index
-                vals_in_idx = sum(vals.isin(idx))
-                
-                if len(idx.names) == 1:
-                    idx_str = idx.names[0]
-                else:
-                    idx_str = '[{}]'.format(','.join(idx.names))
-                
-                print("'{}.{}': {} of {} unique values are found in '{}.{}' ({}%)"\
-                        .format(t1, idx_str, 
-                                vals_in_idx, len(vals), 
-                                t2, idx_str, 
-                                round(100*vals_in_idx/len(vals))))
-    
-        return True
-    
-
